+++ client/src/main/java/com/paraguacraft/pvp/ParaguacraftPvP.java
@@ -27,7 +27,7 @@ public class ParaguacraftPvP {
 
     public static final String MODID = "paraguacraftpvp";
     public static final String NAME = "Paraguacraft PvP Client";
-    public static final String VERSION = "2.1.3";
+    public static final String VERSION = "2.1.5";
 
     @Mod.Instance(MODID)
     public static ParaguacraftPvP instance;